+++ Changelog.md
@@ -5,6 +5,25 @@ All notable Changes to the Julia package `Manopt.jl` will be documented in this 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.x] - dd/mm/2023
+
+### Changed
+
+* References in the documentation are now rendered using [DocumenterCitations.jl](https://github.com/JuliaDocs/DocumenterCitations.jl)
+* Asymptote export now also accepts a size in pixel instead of its default `4cm` size and `render` can be deactivated setting it to `nothing`.
+
+## [0.4.29] - 12/07/2023
+
+### Fixed
+
+* fixed a bug, where `cyclic_proximal_point` did not work with decorated objectives.
+
+## [0.4.28] - 24/06/2023
+
+### Changed
+
+* `max_stepsize` was specialized for `FixedRankManifold` to follow Matlab Manopt.
+
 ## [0.4.27] - 15/06/2023
 
 ### Added